# Remove StopException for control flow