- Using exceptions for control flow is very expensive for performance.
  After removing it, it basically doubled in speed.

With my benchmark here, the time went from 4762ms to 2761ms.

    package net.razorvine.pickle.test;

    import java.io.IOException;
    import java.util.ArrayList;
    import java.util.List;

    import net.razorvine.pickle.PickleException;
    import net.razorvine.pickle.Pickler;
    import net.razorvine.pickle.Unpickler;

    import org.junit.Test;

    public class PerfTest {

        @Test
        public void testPerf() throws PickleException, IOException {
            Pickler pickler = new Pickler();

            List<String> myList = new ArrayList<String>();
            for (int i = 0; i < 100; i++) {
                myList.add(String.valueOf(i));
            }

            byte[] bytes = pickler.dumps(myList);

            Unpickler unpickler = new Unpickler();

            long start = System.currentTimeMillis();
            for (int i = 0; i < 100000; i++) {
                unpickler.loads(bytes);
            }

            System.out.println(System.currentTimeMillis() - start);
        }
    }