@irmen 2015-06-24
Are you saying that even though the exception is thrown only once (at the end), removing it doubles the speed?