@justinuang 2015-06-24
I'm using Oracle JDK 1.8 on mac. You're right, you should only get about 3800 -> 3400ms. My measurements was because I was using debug mode. Using non-debug mode, I get about 3100 -> 2700ms.

The main thing is that if I am calling Unpickle many times with small objects, then the speed up is pretty large (I'm remembering to do non-debug mode this time).

I changed the size of the list I'm deserializing to a list of one integer, and now the times are 418 -> 213ms.